+++ vnext/Microsoft.ReactNative.Managed.UnitTests/NoAttributeNativeModuleTest.cs
@@ -122,6 +122,21 @@ namespace Microsoft.ReactNative.Managed.UnitTests
       resolve("Static Hello_2");
     }
 
+    public void CallbackZeroArgs(Action resolve)
+    {
+      resolve();
+    }
+
+    public void CallbackTwoArgs(Action<int, int> resolve)
+    {
+      resolve(1, 2);
+    }
+
+    public void CallbackThreeArgs(Action<int, int, string> resolve)
+    {
+      resolve(1, 2, "Hello");
+    }
+
     public void DivideCallbacks(int x, int y, Action<int> resolve, Action<string> reject)
     {
       if (y != 0)
@@ -215,6 +230,36 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     public static void StaticRejectSayHelloCallbacks(Action<string> _, Action<string> reject)
     {
       reject("Goodbye");
+    }
+
+    public void TwoCallbacksZeroArgs1(Action resolve, Action _)
+    {
+      resolve();
+    }
+
+    public void TwoCallbacksZeroArgs2(Action _, Action reject)
+    {
+      reject();
+    }
+
+    public void TwoCallbacksTwoArgs1(Action<int, int> resolve, Action<int, int> _)
+    {
+      resolve(1, 2);
+    }
+
+    public void TwoCallbacksTwoArgs2(Action<int, int> _, Action<int, int> reject)
+    {
+      reject(1, 2);
+    }
+
+    public void TwoCallbacksThreeArgs1(Action<int, int, string> resolve, Action<int, int, string> _)
+    {
+      resolve(1, 2, "Hello");
+    }
+
+    public void TwoCallbacksThreeArgs2(Action<int, int, string> _, Action<int, int, string> reject)
+    {
+      reject(1, 2, "Hello");
     }
 
     public void DividePromise(int x, int y, IReactPromise<int> promise)
@@ -546,13 +591,43 @@ namespace Microsoft.ReactNative.Managed.UnitTests
         });
     }
 
-    public void AddMethod<TResult>(string name, Func<TModule, Action<Action<TResult>>> getMethod)
+    public void AddMethod(string name, Func<TModule, Action<Action>> getMethod)
     {
       m_moduleBuilder.AddMethod(name, MethodReturnType.Callback,
         (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
         {
           var method = getMethod(m_module);
-          method(result => resolve(outputWriter.WriteArgs(result)));
+          method(() => resolve(outputWriter.WriteArgs()));
+        });
+    }
+
+    public void AddMethod<TResult1>(string name, Func<TModule, Action<Action<TResult1>>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.Callback,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method(result1 => resolve(outputWriter.WriteArgs(result1)));
+        });
+    }
+
+    public void AddMethod<TResult1, TResult2>(string name, Func<TModule, Action<Action<TResult1, TResult2>>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.Callback,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method((result1, result2) => resolve(outputWriter.WriteArgs(result1, result2)));
+        });
+    }
+
+    public void AddMethod<TResult1, TResult2, TResult3>(string name, Func<TModule, Action<Action<TResult1, TResult2, TResult3>>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.Callback,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method((result1, result2, result3) => resolve(outputWriter.WriteArgs(result1, result2, result3)));
         });
     }
 
@@ -578,6 +653,16 @@ namespace Microsoft.ReactNative.Managed.UnitTests
         });
     }
 
+    public void AddMethod(string name, Func<TModule, Action<Action, Action>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.TwoCallbacks,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method(() => resolve(outputWriter.WriteArgs()), () => reject(outputWriter.WriteArgs()));
+        });
+    }
+
     public void AddMethod<T1, T2>(string name, Func<TModule, Action<Action<T1>, Action<T2>>> getMethod)
     {
       m_moduleBuilder.AddMethod(name, MethodReturnType.TwoCallbacks,
@@ -585,6 +670,30 @@ namespace Microsoft.ReactNative.Managed.UnitTests
         {
           var method = getMethod(m_module);
           method(result => resolve(outputWriter.WriteArgs(result)), result => reject(outputWriter.WriteArgs(result)));
+        });
+    }
+
+    public void AddMethod<T1, T2, T3, T4>(string name, Func<TModule, Action<Action<T1, T2>, Action<T3, T4>>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.TwoCallbacks,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method(
+            (result1, result2) => resolve(outputWriter.WriteArgs(result1, result2)),
+            (result1, result2) => reject(outputWriter.WriteArgs(result1, result2)));
+        });
+    }
+
+    public void AddMethod<T1, T2, T3, T4, T5, T6>(string name, Func<TModule, Action<Action<T1, T2, T3>, Action<T4, T5, T6>>> getMethod)
+    {
+      m_moduleBuilder.AddMethod(name, MethodReturnType.TwoCallbacks,
+        (IJSValueReader inputReader, IJSValueWriter outputWriter, MethodResultCallback resolve, MethodResultCallback reject) =>
+        {
+          var method = getMethod(m_module);
+          method(
+            (result1, result2, result3) => resolve(outputWriter.WriteArgs(result1, result2, result3)),
+            (result1, result2, result3) => reject(outputWriter.WriteArgs(result1, result2, result3)));
         });
     }
 
@@ -686,7 +795,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSEvent(string eventEmitterName, string name, Action<TModule, Action> setEventHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setEventHandler(m_module, () => new ReactContext(reactContext).EmitJSEvent(
           eventEmitterName ?? EventEmitterName ?? "RCTDeviceEventEmitter", name));
       });
@@ -694,7 +804,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSEvent<T1>(string eventEmitterName, string name, Action<TModule, Action<T1>> setEventHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setEventHandler(m_module, (T1 arg1) => new ReactContext(reactContext).EmitJSEvent(
           eventEmitterName ?? EventEmitterName ?? "RCTDeviceEventEmitter", name, arg1));
       });
@@ -702,7 +813,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSEvent<T1, T2>(string eventEmitterName, string name, Action<TModule, Action<T1, T2>> setEventHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setEventHandler(m_module, (T1 arg1, T2 arg2) => new ReactContext(reactContext).EmitJSEvent(
           eventEmitterName ?? EventEmitterName ?? "RCTDeviceEventEmitter", name, arg1, arg2));
       });
@@ -710,7 +822,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSFunction(string moduleName, string name, Action<TModule, Action> setFunctionHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setFunctionHandler(m_module, () => new ReactContext(reactContext).CallJSFunction(
           moduleName ?? ModuleName, name));
       });
@@ -718,7 +831,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSFunction<T1>(string moduleName, string name, Action<TModule, Action<T1>> setFunctionHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setFunctionHandler(m_module, (T1 arg1) => new ReactContext(reactContext).CallJSFunction(
           moduleName ?? ModuleName, name, arg1));
       });
@@ -726,7 +840,8 @@ namespace Microsoft.ReactNative.Managed.UnitTests
 
     public void AddJSFunction<T1, T2>(string moduleName, string name, Action<TModule, Action<T1, T2>> setFunctionHandler)
     {
-      m_moduleBuilder.AddInitializer(reactContext => {
+      m_moduleBuilder.AddInitializer(reactContext =>
+      {
         setFunctionHandler(m_module, (T1 arg1, T2 arg2) => new ReactContext(reactContext).CallJSFunction(
           moduleName ?? ModuleName, name, arg1, arg2));
       });
@@ -758,7 +873,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
       mb.ModuleName = "SimpleNativeModule2";
 
       mb.AddInitializer(m => m.Initialize);
-    
+
       mb.AddMethod("Add", m => new Func<int, int, int>(m.Add));
       mb.AddMethod("Negate", m => new Func<int, int>(m.Negate));
       mb.AddMethod("SayHello", m => new Func<string>(m.SayHello));
@@ -782,6 +897,10 @@ namespace Microsoft.ReactNative.Managed.UnitTests
       mb.AddMethod("StaticNegateAsyncCallback", m => new Action<int, Action<int>>(SimpleNativeModule2.StaticNegateAsyncCallback));
       mb.AddMethod("StaticSayHelloCallback", m => new Action<Action<string>>(SimpleNativeModule2.StaticSayHelloCallback));
 
+      mb.AddMethod("CallbackZeroArgs", m => new Action<Action>(m.CallbackZeroArgs));
+      mb.AddMethod("CallbackTwoArgs", m => new Action<Action<int, int>>(m.CallbackTwoArgs));
+      mb.AddMethod("CallbackThreeArgs", m => new Action<Action<int, int, string>>(m.CallbackThreeArgs));
+
       mb.AddMethod("DivideCallbacks", m => new Action<int, int, Action<int>, Action<string>>(m.DivideCallbacks));
       mb.AddMethod("NegateCallbacks", m => new Action<int, Action<int>, Action<string>>(m.NegateCallbacks));
       mb.AddMethod("NegateAsyncCallbacks", m => new Action<int, Action<int>, Action<string>>(m.NegateAsyncCallbacks));
@@ -792,6 +911,13 @@ namespace Microsoft.ReactNative.Managed.UnitTests
       mb.AddMethod("StaticNegateAsyncCallbacks", m => new Action<int, Action<int>, Action<string>>(SimpleNativeModule2.StaticNegateAsyncCallbacks));
       mb.AddMethod("StaticResolveSayHelloCallbacks", m => new Action<Action<string>, Action<string>>(SimpleNativeModule2.StaticResolveSayHelloCallbacks));
       mb.AddMethod("StaticRejectSayHelloCallbacks", m => new Action<Action<string>, Action<string>>(SimpleNativeModule2.StaticRejectSayHelloCallbacks));
+
+      mb.AddMethod("TwoCallbacksZeroArgs1", m => new Action<Action, Action>(m.TwoCallbacksZeroArgs1));
+      mb.AddMethod("TwoCallbacksZeroArgs2", m => new Action<Action, Action>(m.TwoCallbacksZeroArgs2));
+      mb.AddMethod("TwoCallbacksTwoArgs1", m => new Action<Action<int, int>, Action<int, int>>(m.TwoCallbacksTwoArgs1));
+      mb.AddMethod("TwoCallbacksTwoArgs2", m => new Action<Action<int, int>, Action<int, int>>(m.TwoCallbacksTwoArgs2));
+      mb.AddMethod("TwoCallbacksThreeArgs1", m => new Action<Action<int, int, string>, Action<int, int, string>>(m.TwoCallbacksThreeArgs1));
+      mb.AddMethod("TwoCallbacksThreeArgs2", m => new Action<Action<int, int, string>, Action<int, int, string>>(m.TwoCallbacksThreeArgs2));
 
       mb.AddMethod("DividePromise", m => new Action<int, int, IReactPromise<int>>(m.DividePromise));
       mb.AddMethod("NegatePromise", m => new Action<int, IReactPromise<int>>(m.NegatePromise));
@@ -857,42 +983,42 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_Add()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.Add), 3, 5, (int result) => Assert.AreEqual(8, result));
+      m_moduleBuilderMock.Call1<int, int, Action<int>>(nameof(SimpleNativeModule2.Add), 3, 5, (int result) => Assert.AreEqual(8, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_Negate()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.Negate), 3, (int result) => Assert.AreEqual(-3, result));
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.Negate), 3, (int result) => Assert.AreEqual(-3, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_SayHello()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.SayHello), (string result) => Assert.AreEqual("Hello", result));
+      m_moduleBuilderMock.Call1<Action<string>>(nameof(SimpleNativeModule2.SayHello), (string result) => Assert.AreEqual("Hello", result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticAdd()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticAdd), 3, 5, (int result) => Assert.AreEqual(8, result));
+      m_moduleBuilderMock.Call1<int, int, Action<int>>(nameof(SimpleNativeModule2.StaticAdd), 3, 5, (int result) => Assert.AreEqual(8, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticNegate()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticNegate), 3, (int result) => Assert.AreEqual(-3, result));
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.StaticNegate), 3, (int result) => Assert.AreEqual(-3, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticSayHello()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticSayHello), (string result) => Assert.AreEqual("Hello", result));
+      m_moduleBuilderMock.Call1<Action<string>>(nameof(SimpleNativeModule2.StaticSayHello), (string result) => Assert.AreEqual("Hello", result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
@@ -942,63 +1068,93 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_AddCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.AddCallback), 7, -8, (int result) => Assert.AreEqual(-1, result));
+      m_moduleBuilderMock.Call1<int, int, Action<int>>(nameof(SimpleNativeModule2.AddCallback), 7, -8, (int result) => Assert.AreEqual(-1, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_NegateCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.NegateCallback), 3, (int result) => Assert.AreEqual(-3, result));
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.NegateCallback), 3, (int result) => Assert.AreEqual(-3, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
-    public void TestMethodCall_NegateAyncCallback()
+    public void TestMethodCall_NegateAsyncCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.NegateAsyncCallback), 3, (int result) => Assert.AreEqual(-3, result)).Wait();
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.NegateAsyncCallback), 3, (int result) => Assert.AreEqual(-3, result)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_SayHelloCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.SayHelloCallback), (string result) => Assert.AreEqual("Hello_2", result));
+      m_moduleBuilderMock.Call1<Action<string>>(nameof(SimpleNativeModule2.SayHelloCallback), (string result) => Assert.AreEqual("Hello_2", result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticAddCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticAddCallback), 4, 56, (int result) => Assert.AreEqual(60, result));
+      m_moduleBuilderMock.Call1<int, int, Action<int>>(nameof(SimpleNativeModule2.StaticAddCallback), 4, 56, (int result) => Assert.AreEqual(60, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticNegateCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticNegateCallback), 33, (int result) => Assert.AreEqual(-33, result));
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.StaticNegateCallback), 33, (int result) => Assert.AreEqual(-33, result));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticNegateAsyncCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticNegateAsyncCallback), 33, (int result) => Assert.AreEqual(-33, result)).Wait();
+      m_moduleBuilderMock.Call1<int, Action<int>>(nameof(SimpleNativeModule2.StaticNegateAsyncCallback), 33, (int result) => Assert.AreEqual(-33, result)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_StaticSayHelloCallback()
     {
-      m_moduleBuilderMock.Call1(nameof(SimpleNativeModule2.StaticSayHelloCallback), (string result) => Assert.AreEqual("Static Hello_2", result));
+      m_moduleBuilderMock.Call1<Action<string>>(nameof(SimpleNativeModule2.StaticSayHelloCallback), (string result) => Assert.AreEqual("Static Hello_2", result));
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_CallbackZeroArgs()
+    {
+      m_moduleBuilderMock.Call1<Action>(nameof(SimpleNativeModule.CallbackZeroArgs), () => { });
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_CallbackTwoArgs()
+    {
+      m_moduleBuilderMock.Call1<Action<int, int>>(nameof(SimpleNativeModule.CallbackTwoArgs), (int p1, int p2) =>
+      {
+        Assert.AreEqual(1, p1);
+        Assert.AreEqual(2, p2);
+      });
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_CallbackThreeArgs()
+    {
+      m_moduleBuilderMock.Call1<Action<int, int, string>>(nameof(SimpleNativeModule.CallbackThreeArgs), (int p1, int p2, string p3) =>
+      {
+        Assert.AreEqual(1, p1);
+        Assert.AreEqual(2, p2);
+        Assert.AreEqual("Hello", p3);
+      });
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
     }
 
     [TestMethod]
     public void TestMethodCall_DivideCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.DivideCallbacks), 6, 2,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.DivideCallbacks), 6, 2,
           (int result) => Assert.AreEqual(3, result),
           (string error) => Assert.AreEqual("Division by 0", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1007,7 +1163,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_DivideCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.DivideCallbacks), 6, 0,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.DivideCallbacks), 6, 0,
           (int result) => Assert.AreEqual(3, result),
           (string error) => Assert.AreEqual("Division by 0", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1016,7 +1172,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateCallbacks), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.NegateCallbacks), 5,
           (int result) => Assert.AreEqual(-5, result),
           (string error) => Assert.AreEqual("Already negative", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1025,7 +1181,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateCallbacks), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.NegateCallbacks), -5,
           (int result) => Assert.AreEqual(5, result),
           (string error) => Assert.AreEqual("Already negative", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1034,7 +1190,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateAsyncCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateAsyncCallbacks), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.NegateAsyncCallbacks), 5,
           (int result) => Assert.AreEqual(-5, result),
           (string error) => Assert.AreEqual("Already negative", error)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1043,7 +1199,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateAsyncCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateAsyncCallbacks), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.NegateAsyncCallbacks), -5,
           (int result) => Assert.AreEqual(5, result),
           (string error) => Assert.AreEqual("Already negative", error)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1052,7 +1208,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_ResolveSayHelloCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.ResolveSayHelloCallbacks),
+      m_moduleBuilderMock.Call2<Action<string>, Action<string>>(nameof(SimpleNativeModule2.ResolveSayHelloCallbacks),
           (string result) => Assert.AreEqual("Hello_3", result),
           (string error) => Assert.AreEqual("Goodbye", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1061,7 +1217,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_RejectSayHelloCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.RejectSayHelloCallbacks),
+      m_moduleBuilderMock.Call2<Action<string>, Action<string>>(nameof(SimpleNativeModule2.RejectSayHelloCallbacks),
           (string result) => Assert.AreEqual("Hello_3", result),
           (string error) => Assert.AreEqual("Goodbye", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1070,7 +1226,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticDivideCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticDivideCallbacks), 6, 2,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticDivideCallbacks), 6, 2,
           (int result) => Assert.AreEqual(3, result),
           (string error) => Assert.AreEqual("Division by 0", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1079,7 +1235,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticDivideCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticDivideCallbacks), 6, 0,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticDivideCallbacks), 6, 0,
           (int result) => Assert.AreEqual(3, result),
           (string error) => Assert.AreEqual("Division by 0", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1088,7 +1244,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateCallbacks), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticNegateCallbacks), 5,
           (int result) => Assert.AreEqual(-5, result),
           (string error) => Assert.AreEqual("Already negative", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1097,7 +1253,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateCallbacks), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticNegateCallbacks), -5,
           (int result) => Assert.AreEqual(5, result),
           (string error) => Assert.AreEqual("Already negative", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1106,7 +1262,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateAsyncCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateAsyncCallbacks), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticNegateAsyncCallbacks), 5,
           (int result) => Assert.AreEqual(-5, result),
           (string error) => Assert.AreEqual("Already negative", error)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1115,7 +1271,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateAsyncCallbacksError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateAsyncCallbacks), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<string>>(nameof(SimpleNativeModule2.StaticNegateAsyncCallbacks), -5,
           (int result) => Assert.AreEqual(5, result),
           (string error) => Assert.AreEqual("Already negative", error)).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1124,7 +1280,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticResolveSayHelloCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticResolveSayHelloCallbacks),
+      m_moduleBuilderMock.Call2<Action<string>, Action<string>>(nameof(SimpleNativeModule2.StaticResolveSayHelloCallbacks),
           (string result) => Assert.AreEqual("Hello_3", result),
           (string error) => Assert.AreEqual("Goodbye", error));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1133,16 +1289,90 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticRejectSayHelloCallbacks()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticRejectSayHelloCallbacks),
+      m_moduleBuilderMock.Call2<Action<string>, Action<string>>(nameof(SimpleNativeModule2.StaticRejectSayHelloCallbacks),
           (string result) => Assert.AreEqual("Hello_3", result),
           (string error) => Assert.AreEqual("Goodbye", error));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
     }
 
     [TestMethod]
+    public void TestMethodCall_TwoCallbacksZeroArgs1()
+    {
+      m_moduleBuilderMock.Call2<Action, Action>(nameof(SimpleNativeModule.TwoCallbacksZeroArgs1),
+        () => { }, () => { Assert.Fail(); });
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_TwoCallbacksZeroArgs2()
+    {
+      m_moduleBuilderMock.Call2<Action, Action>(nameof(SimpleNativeModule.TwoCallbacksZeroArgs2),
+        () => { Assert.Fail(); }, () => { });
+      Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_TwoCallbacksTwoArgs1()
+    {
+      m_moduleBuilderMock.Call2<Action<int, int>, Action<int, int>>(
+        nameof(SimpleNativeModule.TwoCallbacksTwoArgs1),
+        (int p1, int p2) =>
+        {
+          Assert.AreEqual(1, p1);
+          Assert.AreEqual(2, p2);
+        },
+        (int p1, int p2) => { Assert.Fail(); });
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_TwoCallbacksTwoArgs2()
+    {
+      m_moduleBuilderMock.Call2<Action<int, int>, Action<int, int>>(
+        nameof(SimpleNativeModule.TwoCallbacksTwoArgs2),
+        (int p1, int p2) => { Assert.Fail(); },
+        (int p1, int p2) =>
+        {
+          Assert.AreEqual(1, p1);
+          Assert.AreEqual(2, p2);
+        });
+      Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_TwoCallbacksThreeArgs1()
+    {
+      m_moduleBuilderMock.Call2<Action<int, int, string>, Action<int, int, string>>(
+        nameof(SimpleNativeModule.TwoCallbacksThreeArgs1),
+        (int p1, int p2, string p3) =>
+        {
+          Assert.AreEqual(1, p1);
+          Assert.AreEqual(2, p2);
+          Assert.AreEqual("Hello", p3);
+        },
+        (int p1, int p2, string p3) => { Assert.Fail(); });
+      Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
+    }
+
+    [TestMethod]
+    public void TestMethodCall_TwoCallbacksThreeArgs2()
+    {
+      m_moduleBuilderMock.Call2<Action<int, int, string>, Action<int, int, string>>(
+        nameof(SimpleNativeModule.TwoCallbacksThreeArgs2),
+        (int p1, int p2, string p3) => { Assert.Fail(); },
+        (int p1, int p2, string p3) =>
+        {
+          Assert.AreEqual(1, p1);
+          Assert.AreEqual(2, p2);
+          Assert.AreEqual("Hello", p3);
+        });
+      Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
+    }
+
+    [TestMethod]
     public void TestMethodCall_DividePromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.DividePromise), 6, 2,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.DividePromise), 6, 2,
           (int result) => Assert.AreEqual(3, result),
           (JSValue error) => Assert.AreEqual("Division by 0", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1151,7 +1381,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_DividePromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.DividePromise), 6, 0,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.DividePromise), 6, 0,
           (int result) => Assert.AreEqual(3, result),
           (JSValue error) => Assert.AreEqual("Division by 0", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1160,7 +1390,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegatePromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegatePromise), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.NegatePromise), 5,
           (int result) => Assert.AreEqual(-5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1169,7 +1399,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegatePromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegatePromise), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.NegatePromise), -5,
           (int result) => Assert.AreEqual(5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1178,7 +1408,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateAsyncPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateAsyncPromise), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.NegateAsyncPromise), 5,
           (int result) => Assert.AreEqual(-5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"])).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1187,7 +1417,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_NegateAsyncPromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.NegateAsyncPromise), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.NegateAsyncPromise), -5,
           (int result) => Assert.AreEqual(5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"])).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1196,7 +1426,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_VoidPromise()
     {
-      m_moduleBuilderMock.Call2("voidPromise", 2,
+      m_moduleBuilderMock.Call2<int, Action<JSValue.Void>, Action<JSValue>>("voidPromise", 2,
           (JSValue.Void result) => { },
           (JSValue error) => Assert.AreEqual("Odd unexpected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1205,7 +1435,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_VoidPromiseError()
     {
-      m_moduleBuilderMock.Call2("voidPromise", 3,
+      m_moduleBuilderMock.Call2<int, Action<JSValue.Void>, Action<JSValue>>("voidPromise", 3,
           (JSValue.Void result) => { },
           (JSValue error) => Assert.AreEqual("Odd unexpected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1214,7 +1444,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_ResolveSayHelloPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.ResolveSayHelloPromise),
+      m_moduleBuilderMock.Call2<Action<string>, Action<JSValue>>(nameof(SimpleNativeModule2.ResolveSayHelloPromise),
           (string result) => Assert.AreEqual("Hello_3", result),
           (JSValue error) => Assert.AreEqual("Promise rejected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1223,7 +1453,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_RejectSayHelloPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.RejectSayHelloPromise),
+      m_moduleBuilderMock.Call2<Action<string>, Action<JSValue>>(nameof(SimpleNativeModule2.RejectSayHelloPromise),
           (string result) => Assert.AreEqual("Hello_3", result),
           (JSValue error) => Assert.AreEqual("Promise rejected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1232,7 +1462,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticDividePromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticDividePromise), 6, 2,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticDividePromise), 6, 2,
           (int result) => Assert.AreEqual(3, result),
           (JSValue error) => Assert.AreEqual("Division by 0", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1241,7 +1471,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticDividePromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticDividePromise), 6, 0,
+      m_moduleBuilderMock.Call2<int, int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticDividePromise), 6, 0,
           (int result) => Assert.AreEqual(3, result),
           (JSValue error) => Assert.AreEqual("Division by 0", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1250,7 +1480,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegatePromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegatePromise), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticNegatePromise), 5,
           (int result) => Assert.AreEqual(-5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1259,7 +1489,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegatePromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegatePromise), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticNegatePromise), -5,
           (int result) => Assert.AreEqual(5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1268,7 +1498,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateAsyncPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateAsyncPromise), 5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticNegateAsyncPromise), 5,
           (int result) => Assert.AreEqual(-5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"])).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1277,7 +1507,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticNegateAsyncPromiseError()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticNegateAsyncPromise), -5,
+      m_moduleBuilderMock.Call2<int, Action<int>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticNegateAsyncPromise), -5,
           (int result) => Assert.AreEqual(5, result),
           (JSValue error) => Assert.AreEqual("Already negative", error["message"])).Wait();
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1286,7 +1516,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticVoidPromise()
     {
-      m_moduleBuilderMock.Call2("staticVoidPromise", 2,
+      m_moduleBuilderMock.Call2<int, Action<JSValue.Void>, Action<JSValue>>("staticVoidPromise", 2,
           (JSValue.Void result) => { },
           (JSValue error) => Assert.AreEqual("Odd unexpected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1295,7 +1525,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticVoidPromiseError()
     {
-      m_moduleBuilderMock.Call2("staticVoidPromise", 3,
+      m_moduleBuilderMock.Call2<int, Action<JSValue.Void>, Action<JSValue>>("staticVoidPromise", 3,
           (JSValue.Void result) => { },
           (JSValue error) => Assert.AreEqual("Odd unexpected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);
@@ -1305,7 +1535,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticResolveSayHelloPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticResolveSayHelloPromise),
+      m_moduleBuilderMock.Call2<Action<string>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticResolveSayHelloPromise),
           (string result) => Assert.AreEqual("Hello_3", result),
           (JSValue error) => Assert.AreEqual("Promise rejected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsResolveCallbackCalled);
@@ -1314,7 +1544,7 @@ namespace Microsoft.ReactNative.Managed.UnitTests
     [TestMethod]
     public void TestMethodCall_StaticRejectSayHelloPromise()
     {
-      m_moduleBuilderMock.Call2(nameof(SimpleNativeModule2.StaticRejectSayHelloPromise),
+      m_moduleBuilderMock.Call2<Action<string>, Action<JSValue>>(nameof(SimpleNativeModule2.StaticRejectSayHelloPromise),
           (string result) => Assert.AreEqual("Hello_3", result),
           (JSValue error) => Assert.AreEqual("Promise rejected", error["message"]));
       Assert.IsTrue(m_moduleBuilderMock.IsRejectCallbackCalled);